ISSUE-10 : completes moving core application endpoints GET /ready/, GET /health/ and GET /version/ to toolbox library
